python docs: fix some of the examples that were not running (#206)
* fixedpoints: remove commented-out code

* programming docs: in the python guide, use print() function instead of print statement

The examples would not run on a recent python, where print() is a function.

Command used to identify places in need to be fixed:
    rg "print\s[^\(]"

* programming docs: fix erroneous indentation in fixedpoints, which broke the Bakery example

The Bakery example would not run because of wrong indentation. After fixing the
indentation it now runs as intended.

For uniformity, the commit also removes tabs from the indentation of the
"Functional Programs" example, which was already working, anyways.

The "Explanations" and "Procedure Calls" were broken before this commit. They
are not affected by these changes, and stay broken as they were.

---------

Co-authored-by: muxator <antonio.muci@bancaditalia.it>