ci: extract z3py examples from the documentation and execute them (#211)
ci: script that extracts z3py examples from the documentation and executes them

This commit introduces a simple CI check for the python section of the z3 guide.

The pipeline is written in python, and works as follows:

- it enumerates all the `.md` files in
  `website/docs-programming/02 - Z3 Python - Readonly`;
- for each of them, it looks for code sections, selecting those whose syntax
  highlighting is set as `z3-python` (let's call them z3 snippets), and ignoring
  the others;
- since each z3 snippet implicitly assumes that the user has imported the whole
  z3 namespace, it prepends a `from z3 import *` statement to each snippet;
- the snippet is then executed in a
  [Python 3.14 subinterpreter](https://docs.python.org/3/library/concurrent.interpreters.htm)
  (so that no new process needs to be spawned, and the global namespace is kept
  pristine), capturing its `stdout` and `stderr`, while catching eventual
  errors.

Each z3 snippet is checked against the latest version of the
[z3-solver package](https://pypi.org/project/z3-solver/) published in PyPA at
that moment, i.e. the one that would be installed executing
`pip install z3-solver` without further version specifications.

The results are then printed in Github flavoured Markdown format on stdout, and
assigned to $GITHUB_STEP_SUMMARY. This results in a test report in the workflow
window.

In the current version, a test on a z3 snippet is considered successful if its
execution merely terminates without exceptions or syntax errors. No check is
performed on the expected output yet.

A stricter check can be implemented later, but I think it would be important
reaching a consensus on how to specify the expected output (ideally in the md
itself?). Moreover, I do not know if all the snippets have deterministic
outputs.

Preliminary discussion at: https://github.com/microsoft/z3guide/pull/206#issuecomment-3311465758

Co-authored-by: muxator <antonio.muci@bancaditalia.it>