ci: add suport for "--html" option when checking z3py documentation (for local development) (#215)
* ci: sort imports and minor reformats in check-z3py-docs

This silences some warnings from ruff.

No functional changes.

* ci: in execute_z3py_docs, refactor the invocation options for main() in a OutputMode enum

This is a preparatory change for the next commit.

No functional changes.

* ci: add support for "--html" flag for local development

This commit introduces support for the "--html" option. When "--html" is given,
the execution report is not written on stdout in Markdown. It is instead
rendered as html in a temporary file and shown in the local web browser.

This change possibly speeds up the cycle time when working locally, because we
can have more or less the same human-friendly output that the CI pipeline gives
us, without having to trigger a remote CI job. No additional dependencies are
introduced.


The current USAGE screen is now:

```
$ scripts/check-z3py-docs.py --help
usage: check-z3py-docs.py [-h] [-r] [--quiet | --html] [file_or_dir]

Parses a single Markdown file, or a directory of files, extracts z3-python snippets and executes them

positional arguments:
  file_or_dir           An existing file or directory. If it is a directory, it will iterate through all "*.md" files contained within. If not given, it will default to "/home/muxator/code/z3guide/website/docs-programming/02 - Z3
                        Python - Readonly"

options:
  -h, --help            show this help message and exit
  -r, --recursive       if given and ARG is a directory, recursively descend in its subdirectories

Output options:
  None of these options are necessary for usage in the CI pipeline. They are, however, useful for local usage.

  --quiet               do not print anything on stdout. Just show logs on stderr (for local usage)
  --html                do not print on stdout the markdown report. Instead, render it in html and open it in a web browser window (for local usage)
```

---------

Co-authored-by: muxator <antonio.muci@bancaditalia.it>